[FIXED] Moved js files to JS folder in assets
