+++ src/assets/js/slider.js
